move js requires
