+++ index.html
@@ -11,6 +11,6 @@
   </head>  
   <body>
     <div id='map' class='map'></div>
-    <script src='js/osrm.bundle.js'></script>
+    <script src='bundle.js'></script>
   </body>
 </html>
