+++ package.json
@@ -2,13 +2,13 @@
   "name": "osrm-frontend",
   "version": "0.4.0",
   "description": "frontend interface for project osrm",
-  "main": "index.js",
+  "main": "js/index.js",
   "scripts": {
-    "test": "eslint index.js src",
-    "build": "browserify -d index.js -s osrm > js/osrm.bundle.js",
-    "build-print": "browserify -d printing.js -s osrm > js/osrm.printing.dev.js",
-    "start-index": "budo index.js --serve=js/osrm.bundle.js --live -d | bistre",
-    "start-print": "budo printing.js --serve=js/osrm.printing.dev.js --live -d | bistre",
+    "test": "eslint js/index.js src",
+    "build": "browserify -d js/index.js -s osrm > bundle.js",
+    "build-print": "browserify -d js/index.print.js -s osrm > bundle.print.js",
+    "start-index": "budo js/index.js --serve=bundle.js --live -d | bistre",
+    "start-print": "budo js/index.print.js --serve=bundle.print.js --live -d | bistre",
     "start": "npm run build && npm run start-index",
     "prepub": "npm run build && npm run build-print"
   },
