+++ printing.html
@@ -12,6 +12,6 @@
   <body>
     <div id='map' class='map'></div>
     <div id='instructions'></div>
-    <script src="js/osrm.printing.dev.js"></script>
+    <script src="bundle.print.js"></script>
   </body>
 </html>
