+++ src/libs/leaflet-label.js
